requirements.txt
requests>=2.31.0
matplotlib>=3.8.0
numpy>=1.26.0
python-Levenshtein>=0.23.0
pytest>=7.4.0
pytest-mock>=3.12.0
python-dotenv>=1.0.0
rich>=13.7.0
